requirements.txt
pyyaml
librosa
pydub
